add comment for feeds

+++ src/entities/comment/model/use-comment-hook.ts
@@ -1,26 +1,28 @@
 import { useUserContext } from "@/shared/stores";
-import { ObjectCommentInterface } from "@/shared/types";
+import { ObjectCommentInterface, PostCommentInterface } from "@/shared/types";
 
 const useCommentHook = ({
     comment,
     onEditComment,
     onDeleteComent,
 }: {
-    comment: ObjectCommentInterface;
-    onEditComment: ({ commentId, text }: { commentId: number; text: string }) => void;
-    onDeleteComent: ({ commentId }: { commentId: number }) => void;
+    comment: ObjectCommentInterface | PostCommentInterface;
+    onEditComment?: ({ commentId, text }: { commentId: number; text: string }) => void;
+    onDeleteComent?: ({ commentId }: { commentId: number }) => void;
 }) => {
     const { user } = useUserContext();
 
     const handleEditComment = () => {
-        onEditComment({ commentId: comment.id, text: comment.commentText });
+        if (onEditComment) {
+            onEditComment({ commentId: comment.id, text: comment.commentText });
+        }
     };
 
     const handleDeleteComment = () => {
-        onDeleteComent({ commentId: comment.id });
+        if (onDeleteComent) {
+            onDeleteComent({ commentId: comment.id });
+        }
     };
-
-    console.log(user, comment);
 
     return { isMyComment: user && user.userId === comment.author.userId, handleEditComment, handleDeleteComment };
 };
